WIIU: Fix L3/R3 buttons
== DETAILS
After a bisect, the culprit was changing the gamepad interface from
returing a single button (bool) to multiple (int16).

The issue is that the Wii U gamepad (and presumably the Pro controller too)
have more than 16 buttons, which means some buttons get lost. Notably, L3 (18)
and R3 (17).

The solution: use int32 instead of int16.

I did a test build and confirmed that this change restores L3/R3 functionality
with the gamepad. Don't have a pro controller to test, but it should work too.

+++ input/common/hid/device_ds3.c
@@ -327,7 +327,7 @@ static int16_t ds3_get_axis(void *data, unsigned axis)
 
 static const char *ds3_get_name(void *data) { return "Sony DualShock 3"; }
 
-static int16_t ds3_button(void *data, uint16_t joykey)
+static int32_t ds3_button(void *data, uint16_t joykey)
 {
    ds3_instance_t *pad = (ds3_instance_t *)data;
    if (!pad || joykey > 31)
