+++ input/input_driver.h
@@ -267,7 +267,7 @@ struct rarch_joypad_driver
    void *(*init)(void *data);
    bool (*query_pad)(unsigned);
    void (*destroy)(void);
-   int16_t (*button)(unsigned, uint16_t);
+   int32_t (*button)(unsigned, uint16_t);
    int16_t (*state)(rarch_joypad_info_t *joypad_info,
          const struct retro_keybind *binds, unsigned port);
    void (*get_buttons)(unsigned, input_bits_t *);
